Gain matrix determined by short term phase correlation
Slices of the phase through time are taken around each frame, and correlation between the left and right phase vectors gives an r value. The r value is then smoothed with a gaussian kernal, and this gives the gain matrix applied to each STFT component. Separation is quite good but there are a number of glitches (especially a tearing glitch which is probably to do with frame transitions).