Gains set using L-R fourier coeffs. difference
Magnitude of difference in fourier coeffecients is taken and normalised by dividing by the magnitude of the sum of the two. This is then fed into a normal distrobution around zero to get gains (sigma roughly 0.3), smoothed with sigma roughly 100 mS. Very good audio quality compared to previous methods, but could perhaps be improved by varying sigma with frequency.